Add number of players selector
