+++ src/components/Game.tsx
@@ -1,4 +1,6 @@
 import React, { useState } from 'react';
+
+import useReactRouter from 'use-react-router';
 import { FiveHundredGame, FiveHundredCard, FiveHundredCardCompare } from '../lib/fivehundred';
 import { Card } from './Card';
 
@@ -8,20 +10,20 @@ interface GameProps {
 }
 
 export const Game: React.SFC<GameProps> = (props) => {
+    const { history } = useReactRouter();
     const [player, setPlayer] = useState(0);
     const [kittyShown, setKittyShown] = useState(false);
     const [hand, setHand] = useState<FiveHundredCard[]>([]);
-
+    const reset = () => {
+        setHand([]);
+        setKittyShown(false);
+    }
 
     const game = new FiveHundredGame(props.players, props.code);
 
     let options = [];
     for (let i=0; i < props.players; i++) {
         options.push(<option key={`optionPlayer${i+1}`} value={i+1}>Player {i+1}</option>);
-    }
-    const reset = () => {
-        setHand([]);
-        setKittyShown(false);
     }
     const handleChange = (event: React.FormEvent<HTMLSelectElement>) => {
         setPlayer(parseInt(event.currentTarget.value));
@@ -38,20 +40,32 @@ export const Game: React.SFC<GameProps> = (props) => {
             setKittyShown(true);
         }
     }
+    const newCode = () => {
+        reset();
+        const randomCode = Math.random().toString(36).slice(2);
+        history.push(`/game/${randomCode}`);
+    }
     return (
         <div>
             <div>
-                <label htmlFor="player-option">I am:</label>
-                <select value={player} onChange={handleChange}>
-                    <option value={0} />
-                    {options}
-                </select>
-                <button disabled={player === 0} onClick={deal}>🃋 Deal me the cards</button>
-                <button disabled={player === 0 || kittyShown} onClick={kitty}>😻 meow</button>
+                <label htmlFor="seed">The secret code for this round is:</label>
+                <input id="seed" readOnly value={props.code}/>
+                <button onClick={newCode}>🎲 New Code</button>
             </div>
-            <ul id="cards">{hand.map((card) => {
-                return <Card key={card.repr()} rank={card.value} suit={card.suit} />
-            })}</ul>
+            <div>
+                <div>
+                    <label htmlFor="player-option">I am:</label>
+                    <select value={player} onChange={handleChange}>
+                        <option value={0} />
+                        {options}
+                    </select>
+                    <button disabled={player === 0} onClick={deal}>🃋 Deal me the cards</button>
+                    <button disabled={player === 0 || kittyShown} onClick={kitty}>😻 meow</button>
+                </div>
+                <ul id="cards">{hand.map((card) => {
+                    return <Card key={card.repr()} rank={card.value} suit={card.suit} />
+                })}</ul>
+            </div>
         </div>
     );
 }
